+++ testing/testing_zgesv_nopiv.c
@@ -124,8 +124,8 @@ testing_zgesv_nopiv_std( run_arg_list_t *args, int check )
     CHAMELEON_Set( CHAMELEON_TILE_SIZE, nb );
 
     /* Creates the matrices */
-    A = malloc( LDA*N*   sizeof(CHAMELEON_Complex64_t) );
-    X = malloc( LDB*NRHS*sizeof(CHAMELEON_Complex64_t) );
+    A = malloc( (size_t) LDA*N*   sizeof(CHAMELEON_Complex64_t) );
+    X = malloc( (size_t) LDB*NRHS*sizeof(CHAMELEON_Complex64_t) );
 
     /* Fills the matrix with random values */
     CHAMELEON_zplgtr( 0,    ChamUpper, N, N, A, LDA, seedA   );
@@ -142,8 +142,8 @@ testing_zgesv_nopiv_std( run_arg_list_t *args, int check )
     if ( check ) {
         CHAMELEON_Complex64_t *A0, *B;
 
-        A0 = malloc( LDA*N   *sizeof(CHAMELEON_Complex64_t) );
-        B  = malloc( LDB*NRHS*sizeof(CHAMELEON_Complex64_t) );
+        A0 = malloc( (size_t) LDA*N   *sizeof(CHAMELEON_Complex64_t) );
+        B  = malloc( (size_t) LDB*NRHS*sizeof(CHAMELEON_Complex64_t) );
         CHAMELEON_zplgtr( 0,    ChamUpper, N, N, A0, LDA, seedA   );
         CHAMELEON_zplgtr( bump, ChamLower, N, N, A0, LDA, seedA+1 );
         CHAMELEON_zplrnt( N, NRHS, B,  LDB, seedB );
